% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vomodel.R
\name{d_enthalpy}
\alias{d_enthalpy}
\title{Partial Voorn-Overbeek mixing enthalpy}
\usage{
d_enthalpy(phi, temp, lattice, sigma)
}
\arguments{
\item{phi}{A numeric vector of length 5. Volume fraction of A~E five species.}

\item{temp}{A number. Temperature in Kelvin}

\item{lattice}{A number. Size of lattice.}

\item{sigma}{A numeric vector of length 5. Charge per monomer of five species.}
}
\description{
\eqn{\partial\DeltaH / \partial x} where \eqn{x = \phi_A}
}
