% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vomodel.R
\name{d_enthalpy_chi}
\alias{d_enthalpy_chi}
\title{Partial Flory-Huggins \Chi enthalpy}
\usage{
d_enthalpy_chi(phi, chi)
}
\arguments{
\item{phi}{A numeric vector of length 5. Volume fraction of A~E five species.}

\item{chi}{A number. Chi value}
}
\description{
Partial Flory-Huggins \Chi enthalpy
}
