% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vomodel.R
\name{d_entropy}
\alias{d_entropy}
\title{Partial ideal mixing entropy}
\usage{
d_entropy(phi, N)
}
\arguments{
\item{phi}{A numeric vector of length 5. Volume fraction of A~E five species.}

\item{N}{An integer vector of length 5. Degree of polymerization of A~E five species.}
}
\description{
\eqn{\partial\DeltaS / \partial x} where \eqn{x = \phi_A}
}
