% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vomodel.R
\name{d_gibbs}
\alias{d_gibbs}
\title{Partial mixing free energy}
\usage{
d_gibbs(phi, temp, N, lattice, sigma, chi)
}
\arguments{
\item{phi}{A numeric vector of length 5. Volume fraction of A~E five species.}

\item{temp}{A number. Temperature in Kelvin}

\item{N}{An integer vector of length 5. Degree of polymerization of A~E five species.}

\item{lattice}{A number. Size of lattice.}

\item{sigma}{A numeric vector of length 5. Charge per monomer of five species.}

\item{chi}{A number. Chi value}
}
\description{
Partial mixing free energy
}
