% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.chi.from.experiment.R,
%   R/get.phase.diagram.using.chitempfunction.r,
%   R/get.phase.diagram.using.chitempfunction.temp.phi3.r
\name{fn}
\alias{fn}
\title{Phase separation function used for root finding}
\usage{
fn(x, phi1, temp, N, lattice, sigma, p2p1, chitemp)

fn(x, phi1, temp, N, lattice, sigma, p2p1, chitemp)

fn(x, phi1, temp, N, lattice, sigma, p2p1, chitemp)
}
\arguments{
\item{phi1}{A number. volume fraction of A in dilute phase}

\item{temp}{A number. Temperature in Kelvin}

\item{N}{An integer vector of length 5. Degree of polymerization of A~E five species.}

\item{lattice}{A number. Size of lattice.}

\item{sigma}{A numeric vector of length 5. Charge per monomer of five species.}

\item{p2p1}{A number. the volume fraction ratio between specie 1 and specie 2}
}
\value{
If inputs are valid, the output will be a length-one numeric vector
}
\description{
\eqn{fn(x) = \frac{\partial G}{\partial x} \times (x - x0) - (G(x0) - G(x))}
}
