% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vomodel.R
\name{get_alpha}
\alias{get_alpha}
\title{Voorn-Overbeek alpha}
\usage{
get_alpha(temp, lattice)
}
\arguments{
\item{temp}{A number. Temperature in Kelvin}

\item{lattice}{A number. Size of lattice.}
}
\description{
Voorn-Overbeek alpha
}
