% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.chi.from.experiment.R
\name{get_chi}
\alias{get_chi}
\title{Get chi value from given constrains}
\usage{
get_chi(phi1, phi2, phi3, temp, N, lattice, sigma, p2p1)
}
\arguments{
\item{phi1}{A number. volume fraction of A in dilute phase}

\item{phi2}{A number. volume fractino of A in dense phase}

\item{phi3}{A number. volume fraction of C monovalent cation}

\item{temp}{A number. Temperature in Kelvin}

\item{N}{An integer vector of length 5. Degree of polymerization of A~E five species.}

\item{lattice}{A number. Size of lattice.}

\item{sigma}{A numeric vector of length 5. Charge per monomer of five species.}

\item{p2p1}{A number. the volume fraction ratio between specie 1 and specie 2}
}
\value{
If inputs are valid, the output will be a length-one numeric vector.
}
\description{
A, B, C, D, E are five species defined in README
}
