% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vomodel.R
\name{sigma.cr}
\alias{sigma.cr}
\title{Counter-ion Release charge density}
\usage{
\method{sigma}{cr}(sigma0, temp)
}
\arguments{
\item{sigma0}{A number. Ideal charge density.}

\item{temp}{A number. Temperature in Kelvin}
}
\value{
A number.
}
\description{
Manning's counter-ion release
}
